# [Perf] Support L3_LOGT_WRITE -> l3_log_write() for benchmarking.