--- a/include/l3.h
+++ b/include/l3.h
@@ -55,6 +55,7 @@ typedef enum {
       L3_LOG_UNDEF      = 0
     , L3_LOG_MMAP
     , L3_LOG_FPRINTF
+    , L3_LOG_WRITE
     , L3_LOGTYPE_MAX
     , L3_LOG_DEFAULT    = L3_LOG_MMAP
 } l3_log_t;
@@ -96,6 +97,10 @@ const char *l3_logtype_name(l3_log_t logtype);
 
     #  define l3_log(msg, arg1, arg2) l3_log_fprintf((msg), arg1, arg2)
 
+    #elif defined(L3_LOGT_WRITE)
+
+    #  define l3_log(msg, arg1, arg2) l3_log_write((msg), arg1, arg2)
+
     #else
 
     #  define l3_log(msg, arg1, arg2)                                   \
@@ -125,6 +130,10 @@ const char *l3_logtype_name(l3_log_t logtype);
     #if defined(L3_LOGT_FPRINTF)
 
     #define l3_log(msg, arg1, arg2) l3_log_fprintf((msg), arg1, arg2)
+
+    #elif defined(L3_LOGT_WRITE)
+
+    #define l3_log(msg, arg1, arg2) l3_log_write((msg), arg1, arg2)
 
     #else
     #define l3_log(msg, arg1, arg2)                                     \
@@ -170,6 +179,8 @@ l3_log_fprintf(const char *msg, const uint64_t arg1, const uint64_t arg2)
 {
     fprintf(l3_log_fh, msg, arg1, arg2);
 }
+
+void l3_log_write(const char *msg, const uint64_t arg1, const uint64_t arg2);
 
 #ifdef __cplusplus
 }