--- a/test.sh
+++ b/test.sh
@@ -101,6 +101,7 @@ TestList=(
 
            "test-build-and-run-client-server-perf-test"
            "test-build-and-run-client-server-perf-test-fprintf"
+           "test-build-and-run-client-server-perf-test-write"
            "test-build-and-run-client-server-perf-test-l3_loc_eq_1"
            "test-build-and-run-client-server-perf-test-l3_loc_eq_2"
 
@@ -437,6 +438,8 @@ function run-all-client-server-perf-tests()
 
     test-build-and-run-client-server-perf-test-fprintf "${num_msgs_per_client}"
 
+    test-build-and-run-client-server-perf-test-write "${num_msgs_per_client}"
+
     test-build-and-run-client-server-perf-test-l3_loc_eq_1 "${num_msgs_per_client}"
 
     test-build-and-run-client-server-perf-test-l3_loc_eq_2 "${num_msgs_per_client}"
@@ -526,14 +529,6 @@ function test-build-and-run-client-server-perf-test()
 
     set +x
     echo " "
-    echo "**** ${Me}: Client-server performance testing with L3-write logging ON:"
-    echo " "
-    build-and-run-client-server-perf-test "${num_msgs_per_client}"  \
-                                           "${l3_log_enabled}"      \
-                                           "${l3_LOC_disabled}"     \
-                                           "write"
-
-    echo " "
     echo "**** ${Me}: Completed basic client(s)-server communication test."
     echo " "
 
@@ -570,6 +565,33 @@ function test-build-and-run-client-server-perf-test-fprintf()
                                            "${l3_log_enabled}"      \
                                            "${l3_LOC_disabled}"     \
                                            "fprintf"
+}
+
+# #############################################################################
+# Test build-and-run of client-server performance test benchmark, using the
+# write() logging interface under L3..
+# #############################################################################
+function test-build-and-run-client-server-perf-test-write()
+{
+    local num_msgs_per_client=${NumMsgsPerClient}
+    if [ $# -ge 1 ]; then
+        num_msgs_per_client=$1
+    fi
+
+    if [ $# -ge 2 ]; then
+        SvrClockArg=$2
+    fi
+
+    local l3_log_enabled=1
+    local l3_LOC_disabled=0
+
+    echo " "
+    echo "**** ${Me}: Client-server performance testing with L3-write logging ON:"
+    echo " "
+    build-and-run-client-server-perf-test "${num_msgs_per_client}"  \
+                                           "${l3_log_enabled}"      \
+                                           "${l3_LOC_disabled}"     \
+                                           "write"
 }
 
 # #############################################################################