--- a/test.sh
+++ b/test.sh
@@ -102,7 +102,6 @@ TestList=(
            "test-build-and-run-client-server-perf-test"
            "test-build-and-run-client-server-perf-test-fprintf"
            "test-build-and-run-client-server-perf-test-write"
-           "test-build-and-run-client-server-perf-test-write-msg"
            "test-build-and-run-client-server-perf-test-l3_loc_eq_1"
            "test-build-and-run-client-server-perf-test-l3_loc_eq_2"
 
@@ -441,8 +440,6 @@ function run-all-client-server-perf-tests()
 
     test-build-and-run-client-server-perf-test-write "${num_msgs_per_client}"
 
-    test-build-and-run-client-server-perf-test-write-msg "${num_msgs_per_client}"
-
     test-build-and-run-client-server-perf-test-l3_loc_eq_1 "${num_msgs_per_client}"
 
     test-build-and-run-client-server-perf-test-l3_loc_eq_2 "${num_msgs_per_client}"
@@ -572,7 +569,7 @@ function test-build-and-run-client-server-perf-test-fprintf()
 
 # #############################################################################
 # Test build-and-run of client-server performance test benchmark, using the
-# write() logging interface under L3..
+# write() msg logging interface under L3.
 # #############################################################################
 function test-build-and-run-client-server-perf-test-write()
 {
@@ -595,33 +592,6 @@ function test-build-and-run-client-server-perf-test-write()
                                            "${l3_log_enabled}"      \
                                            "${l3_LOC_disabled}"     \
                                            "write"
-}
-
-# #############################################################################
-# Test build-and-run of client-server performance test benchmark, using the
-# write()-msg logging interface under L3..
-# #############################################################################
-function test-build-and-run-client-server-perf-test-write-msg()
-{
-    local num_msgs_per_client=${NumMsgsPerClient}
-    if [ $# -ge 1 ]; then
-        num_msgs_per_client=$1
-    fi
-
-    if [ $# -ge 2 ]; then
-        SvrClockArg=$2
-    fi
-
-    local l3_log_enabled=1
-    local l3_LOC_disabled=0
-
-    echo " "
-    echo "**** ${Me}: Client-server performance testing with L3-write logging ON:"
-    echo " "
-    build-and-run-client-server-perf-test "${num_msgs_per_client}"  \
-                                           "${l3_log_enabled}"      \
-                                           "${l3_LOC_disabled}"     \
-                                           "write-msg"
 }
 
 # #############################################################################
@@ -778,16 +748,6 @@ function build-and-run-client-server-perf-test()
                     L3_ENABLED=${l3_enabled}    \
                     BUILD_VERBOSE=1             \
                     L3_LOGT_WRITE=1             \
-                    make client-server-perf-test
-                ;;
-
-            "write-msg")
-                set -x
-                make clean                      \
-                && CC=gcc LD=g++               \
-                    L3_ENABLED=${l3_enabled}    \
-                    BUILD_VERBOSE=1             \
-                    L3_LOGT_WRITE=2             \
                     make client-server-perf-test
                 ;;