--- a/use-cases/client-server-msgs-perf/svmsg_file_server.c
+++ b/use-cases/client-server-msgs-perf/svmsg_file_server.c
@@ -219,6 +219,8 @@ main(int argc, char *argv[])
     logtype = L3_LOG_FPRINTF;
 #elif L3_LOGT_WRITE
     logtype = L3_LOG_WRITE;
+#elif L3_LOGT_WRITE_MSG
+    logtype = L3_LOG_WRITE_MSG;
 #endif
 
     int e = l3_log_init(logtype, logfile);
@@ -232,6 +234,8 @@ main(int argc, char *argv[])
     l3_log_mode = "fprintf() ";
 #elif L3_LOGT_WRITE
     l3_log_mode = "write() ";
+#elif L3_LOGT_WRITE_MSG
+    l3_log_mode = "write()-msg ";
 #else
     l3_log_mode = "";
 #endif  // L3_FASTLOG_ENABLED