… tests/unit/l3-fprintf-perf-test.c
… tests/unit/l3-write-perf-test.c