This commit adds support for `L3_LOGT_WRITE=1 make client-server-perf-test` to build client-server perf-tests to log using write() system call.

- Makefile and #if'def-ed code is added to support write() logging.
- test.sh - Add "write" logging-support to existing test-method.

-----

###   [Perf] Add fprintf/write-logging perf unit-tests

This commit adds couple of new unit-tests to do standalone micro-benchmarking of the performance of l3_log_fprintf(),
 vs l3_log_write() interfaces.

- `make run-unit-tests` will automatically invoke these new tests.
- To resolve usefulness of `write()` v/s writing out a formatted a message, this perf unit-test provides an alternate implementation of `write_msg()` interface thru `l3_log()`.
 
Details in the descriptions below.