@gapisback 2024-05-29
@gregthelaw W.r.t [this comment](https://github.com/undoio/l3/pull/64#pullrequestreview-2085104391) of yours, here is what I reasoned:

As per [this Gnu reference](https://man7.org/linux/man-pages/man3/fwrite.3p.html) on `write()` or `fwrite()`, this writes out binary data:

```
The fwrite() function shall write, from the array pointed to by ptr, up to nitems 
elements whose size is specified by size, to the stream pointed to by stream.
```

fprintf() is useful for format-string processing. 

```
Each of these functions converts, formats, and prints its
       arguments under control of the format. 
```

So you are doing more work at run-time while logging using `fprintf()` v/s `fwrite()`.

----

Re: _As opposed to say using write() to just write the log message directly. I would expect the performance between the two techniques to be indistinguishable, but writing the string directly doesn't require post processing._

I don't think that's entirely true. In order to `write` out the file's name, you would have to pre-format an output buffer @ run-time and `write` that out. That means, L3-logging has to do the work up-front to convert the %s to the file's name etc.

----


Going by [this blog post](https://stackoverflow.com/questions/15079049/write-vs-fprintf-why-different-and-which-is-better), `write` does unbuffered writes to the file, as opposed to `fprintf` or `fwrite`, which may involve file-system buffering.

So, I thought supporting `write` is better (than `fwrite`, which may do file-system buffering) and also useful in that we can write out binary data to the file, just like it's being done with log to `mmap()`-ed file.

With default L3-logging, you are restricted to some n-log-entries. With `write` interface, theoretically the file can keep growing.

By benchmarking against `write`, I'm trying to compare the perf-gains of doing IO to mmap-'ed file, v/s the device IO thru `write` system call; with both schemes needing post-processing (so that part is the same).

Hope this helps.