@gapisback 2024-05-29
@gregthelaw : Fixing the bug in the `write` system call now shows a major difference in overall throughput.

With the bug, we were writing 0-bytes:

```
    **** Performance comparison for NumClients=5, NumOps=5000000 (5 Million) ****
+-------------------------------+-------------------+----------+-------------------+----------+
| Run-Type                      | Server throughput | Srv:Drop | Client throughput | Cli:Drop |
+-------------------------------+-------------------+----------+-------------------+----------+
| Baseline - No logging         |  ~99.19 K ops/sec |  0.00 %  |  ~22.86 K ops/sec |  0.00 %  |

| L3-write() logging (no LOC)   |  ~93.64 K ops/sec | -5.59 %  |  ~21.56 K ops/sec | -5.71 %  |
```

With the bug-fix, here's the new report:

```
    **** Performance comparison for NumClients=5, NumOps=5000000 (5 Million) ****
+-------------------------------+-------------------+----------+-------------------+----------+
| Run-Type                      | Server throughput | Srv:Drop | Client throughput | Cli:Drop |
+-------------------------------+-------------------+----------+-------------------+----------+
| Baseline - No logging         |  ~99.40 K ops/sec |  0.00 %  |  ~24.31 K ops/sec |  0.00 %  |

| L3-write() logging (no LOC)   |  ~77.63 K ops/sec | -21.90 % |  ~16.85 K ops/sec | -30.70 % |
```

Baseline is about the same in both runs ~99.19 K ops/sec.

With the bug-fix, the write-throughput drops from ~93.64 K to ~77.63 K ops/sec, resulting in -21.90 % over baseline.

This seems inline with expected performance of unbuffered write system call.

Do we need to dig more? (The bug-fix is warranted, certainly.)