@gapisback 2024-05-30
@gregthelaw - To address your comment that we should invoke `write()` to log a fully-formatted message, I add unit-tests in this [2nd commit](https://github.com/undoio/l3/pull/64/commits/661a133eaf7d6f91d66125b761bcf87c9117ceb5).

The unit-tests serve multiple purposes:

1. `unit/l3-fprintf-perf-test` exercises L3-fprintf() interfaces added in previous commit. (Test-code coverage.)
2. `unit/l3-write-perf-test` does the same thing for l3-write() interfaces.
3. The `l3-write-perf-test` test exercises both variations of `write` call, to measure if there is any overhead of doing `snprintf()` to create a fully-formed message.